+++ src/apps/hedgefarm/contracts/ethers/index.ts
@@ -2,5 +2,7 @@
 /* tslint:disable */
 /* eslint-disable */
 export type { AlphaOne } from './AlphaOne';
+export type { AlphaTwo } from './AlphaTwo';
 export * as factories from './factories';
 export { AlphaOne__factory } from './factories/AlphaOne__factory';
+export { AlphaTwo__factory } from './factories/AlphaTwo__factory';