+++ src/apps/llama-airforce/ethereum/llama-airforce.vault.token-fetcher.ts
@@ -31,6 +31,7 @@ export class EthereumLlamaAirforceVaultTokenFetcher extends AppTokenTemplatePosi
   getAddresses() {
     return [
       '0x83507cc8c8b67ed48badd1f59f684d5d02884c81', // uCRV
+      '0x4ebad8dbd4edbd74db0278714fbd67ebc76b89b7', // uCRV V2
       '0xf964b0e3ffdea659c44a5a52bc0b82a24b89ce0e', // uFXS
       '0x8659fc767cad6005de79af65dafe4249c57927af', // uCVX
       '0xd6fc1ecd9965ba9cac895654979564a291c74c29', // uauraBAL