+++ src/apps/unstoppable/contracts/index.ts
@@ -4,24 +4,20 @@ import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
 import { ContractFactory } from '~contract/contracts';
 import { Network } from '~types/network.interface';
 
-import { AlphaOne__factory, AlphaTwo__factory } from './ethers';
+import { UnstoppableGlpVault__factory } from './ethers';
 
 // eslint-disable-next-line
 type ContractOpts = { address: string; network: Network };
 
 @Injectable()
-export class HedgefarmContractFactory extends ContractFactory {
+export class UnstoppableContractFactory extends ContractFactory {
   constructor(@Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit) {
     super((network: Network) => appToolkit.getNetworkProvider(network));
   }
 
-  alphaOne({ address, network }: ContractOpts) {
-    return AlphaOne__factory.connect(address, this.appToolkit.getNetworkProvider(network));
-  }
-  alphaTwo({ address, network }: ContractOpts) {
-    return AlphaTwo__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  unstoppableGlpVault({ address, network }: ContractOpts) {
+    return UnstoppableGlpVault__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
 }
 
-export type { AlphaOne } from './ethers';
-export type { AlphaTwo } from './ethers';
+export type { UnstoppableGlpVault } from './ethers';